projects/user_service/app/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
pydantic
PyJWT==2.10.1
alembic==1.16.4
psycopg2-binary>=2.9.0
pydantic[email]
